Refactor calls to `go-gather`
This commit updates calls to the `go-gather` package where applicable.
The majority of the refactor involes updated metadata names and
locations.
